+++ examples/test-site/src/components/Product/index.tsx
@@ -19,22 +19,24 @@ import {
   addClasses,
 } from '@bodiless/fclasses';
 import Tout from '../Tout';
-import { asEditorSimple } from '../Editors';
+import { withEditorSimple } from '../Editors';
 import {
   asHeader1,
   asImage,
   asEditableImage,
   asYMargin,
   asNegXMargin,
+  withPadding5,
 } from '../Elements.token';
 import {
+  asToutWithPaddings,
   asToutDefaultStyle,
   asToutVertical,
 } from '../Tout/token';
 
 export const ProductTitle = flow(
   asHeader1,
-  asEditorSimple('product_title', 'Product Title'),
+  withEditorSimple('product_title', 'Product Title'),
 )(H1);
 
 const asProductImage = addClasses('w-full');
@@ -45,10 +47,12 @@ export const ProductImage = flow(
 )(Img);
 
 export const ProductTout = flow(
+  asToutWithPaddings,
   asToutDefaultStyle,
   asToutVertical,
 )(Tout);
 
+export const SectionContent = withPadding5(Div);
 export const SectionMargin = asYMargin(Div);
 export const SectionNegXMargin = flow(
   asYMargin,