+++ examples/test-site/src/templates/product_detail.jsx
@@ -13,10 +13,12 @@
  */
 
 import React from 'react';
-import { flow } from 'lodash';
+import Helmet from 'react-helmet';
+import { flow, flowRight } from 'lodash';
 import { graphql } from 'gatsby';
 import { Page } from '@bodiless/gatsby-theme-bodiless';
 import { BVRatingsSummary, BVReviews } from '@bodiless/bv';
+import { withNode } from '@bodiless/core';
 import {
   withDesign,
   replaceWith,
@@ -25,6 +27,7 @@ import {
 import {
   SingleAccordionClean,
 } from '@bodiless/organisms';
+import { withEvent, asBodilessHelmet } from '@bodiless/components';
 import Layout from '../components/Layout';
 import {
   ProductTitle,
@@ -34,8 +37,8 @@ import {
   ProductDetailImageWrapper,
   ProductDetailAccWrapper,
 } from '../components/Product';
-import { FlexBoxDefault } from '../components/Flexbox';
-import { asEditorBasic } from '../components/Editors';
+import { FlowContainerDefault } from '../components/FlowContainer';
+import { withEditorBasic } from '../components/Editors';
 import asSingleAccordionDefaultStyle from '../components/SingleAccordion/token';
 
 // Do not allow editors to set accordion titles.
@@ -46,11 +49,12 @@ const NonEditableTitle = ({ producttitle }) => (
 );
 
 const asProductAccordion = title => flow(
+  withNode,
   asSingleAccordionDefaultStyle,
   withDesign({
     Wrapper: removeClasses('p-1'),
     Title: replaceWith(() => <NonEditableTitle producttitle={title} />),
-    Body: asEditorBasic(
+    Body: withEditorBasic(
       'body',
       'Enter Product Information',
     ),
@@ -66,9 +70,16 @@ const ProductInactIngAcc = asProductAccordion('Inactive Ingredients')(SingleAcco
 const ProductStorAcc = asProductAccordion('Storage')(SingleAccordionClean);
 const ProductWarnAcc = asProductAccordion('Warnings')(SingleAccordionClean);
 
+const ExampleGTMHelmetEvent = flowRight(
+  asBodilessHelmet('datalayer'),
+  // On product pages, we may add product related datalayer info:
+  withEvent('digitalData', { event: 'Product Viewed' }, 'product-viewed'),
+)(Helmet);
+
 export default (props: any) => (
   <Page {...props}>
     <Layout>
+      <ExampleGTMHelmetEvent />
       <SectionMargin>
         <div className="flex flex-wrap md:items-end md:flex-row-reverse">
           <div className="w-full md:flex-1 md:flex-grow-0 md:flex-shrink-0 text-right"><p>Placeholder_for_Share</p></div>
@@ -101,7 +112,7 @@ export default (props: any) => (
         <BVReviews />
       </SectionMargin>
       <SectionNegXMargin>
-        <FlexBoxDefault
+        <FlowContainerDefault
           nodeKey="product_touts"
           maxComponents={3}
         />