+++ packages/bodiless-core/src/index.ts
@@ -18,7 +18,12 @@ import PageEditContext from './PageEditContext';
 import asStatic from './asStatic';
 import { useEditContext, useUUID, useContextActivator } from './hooks';
 import withNode, { withNodeKey } from './withNode';
-import withEditButton from './withEditButton';
+import withSidecarNodes, { startSidecarNodes, endSidecarNodes } from './withSidecarNodes';
+import {
+  withDefaultContent,
+  withResetButton,
+} from './Contentful';
+import withEditButton, { EditButtonProps } from './withEditButton';
 import contextMenuForm from './contextMenuForm';
 import withData from './withData';
 import NodeProvider, { useNode, useNodeDataHandlers } from './NodeProvider';
@@ -31,11 +36,13 @@ import {
   withContextActivator,
   withoutProps,
 } from './hoc';
-import { ifEditable, ifReadOnly, withEditToggle } from './withEditToggle';
-import { TMenuOption } from './PageEditContext/types';
+import { ifToggledOff, ifToggledOn, withFlowToggle } from './withFlowToggle';
+import { ifEditable, ifReadOnly, useEditToggle } from './withEditToggle';
+import { TMenuOption, PageEditContextInterface } from './PageEditContext/types';
 import { EditButtonOptions } from './Types/EditButtonTypes';
 import { TMenuOptionGetter } from './Types/PageContextProviderTypes';
 import { WithNodeProps } from './Types/NodeTypes';
+import { TOverlaySettings } from './Types/PageOverlayTypes';
 import {
   ActivateOnEffectProvider,
   withActivateOnEffect,
@@ -51,8 +58,10 @@ export {
   withNodeAndHandlers,
   withNodeDataHandlers,
   withLocalContextMenu,
+  PageEditContextInterface,
   TMenuOption,
   TMenuOptionGetter,
+  TOverlaySettings,
   PageContextProvider as ContextProvider,
   withPageContext as withMenuOptions,
   PageEditContext,
@@ -60,10 +69,14 @@ export {
   useContextActivator,
   useUUID,
   withEditButton,
+  EditButtonProps,
   WithNodeProps,
   EditButtonOptions,
   withNode,
   withNodeKey,
+  withSidecarNodes,
+  startSidecarNodes,
+  endSidecarNodes,
   contextMenuForm,
   withData,
   NodeProvider,
@@ -72,13 +85,18 @@ export {
   DefaultContentNode,
   ifEditable,
   ifReadOnly,
-  withEditToggle,
   withoutProps,
   ActivateOnEffectProvider,
   withActivateOnEffect,
   useActivateOnEffect,
   useActivateOnEffectActivator,
   withChild,
+  withDefaultContent,
+  withResetButton,
+  ifToggledOff,
+  ifToggledOn,
+  withFlowToggle,
+  useEditToggle,
 };
 
 export type Bodiless<P, Q> = (C: ComponentType<P> | string) => ComponentType<Q>;