+++ packages/bodiless-migration-tool/src/scraper.ts
@@ -29,6 +29,7 @@ import HCCrawler = require('@bodiless/headless-chrome-crawler');
 export interface ScrapedPage {
   pageUrl: string,
   rawHtml: string,
+  status: number,
   processedHtml: string,
   metatags: Array<string>,
   scripts: Array<string>,
@@ -42,14 +43,14 @@ export interface ScrapedPage {
 }
 
 interface Events {
-  success(result: ScrapedPage): void,
+  pageReceived(result: ScrapedPage): void,
   fileReceived(file: string): void,
   requestStarted(file: string): void,
   error(error: Error): void
 }
 
 export interface ScraperParams {
-  pageUrl: string,
+  pageUrls: string[],
   maxDepth: number,
   maxConcurrency?: number,
   javascriptEnabled: boolean,
@@ -91,16 +92,25 @@ export class Scraper extends EE<Events> {
       // Function to be called with evaluated results from browsers
       onSuccess: (async successResult => {
         try {
-          const { result } = successResult;
           // we can get an external url here
           // when an internal url is redirected to the external
-          if (isUrlExternal(this.params.pageUrl, successResult.response.url)) {
+          const externalUrl = this.params.pageUrls.some(url => (
+            isUrlExternal(url, successResult.response.url)
+          ));
+          if (externalUrl) {
+            console.log(`external url ${successResult.response.url} received. skipping`);
             return;
           }
-          result.pageUrl = successResult.response.url;
-          // @ts-ignore
-          result.rawHtml = await successResult.rawHtml;
-          this.emit('success', result);
+          // decide if we get page or resource response
+          if (successResult.isHtmlResponse) {
+            const { result, response } = successResult;
+            result.status = response.status;
+            result.pageUrl = successResult.response.url;
+            result.rawHtml = await successResult.responseText;
+            this.emit('pageReceived', result);
+          } else {
+            this.emit('fileReceived', successResult.response.url);
+          }
         } catch (error) {
           debug(error);
         }
@@ -126,17 +136,21 @@ export class Scraper extends EE<Events> {
         this.emit('requestStarted', request.url());
       }
     });
-    const pageHost = getHostNameWithoutWWW(this.params.pageUrl);
-    const allowedDomains = [
-      pageHost,
-      ...pageHost ? [`www.${pageHost}`] : [],
-    ];
     // Queue a request
-    await crawler.queue({
-      url: this.params.pageUrl,
-      maxDepth: this.params.maxDepth,
-      allowedDomains,
+    const queue = this.params.pageUrls.map((url, index, pageUrls) => {
+      const pageHost = getHostNameWithoutWWW(url);
+      const allowedDomains = [
+        pageHost,
+        ...pageHost ? [`www.${pageHost}`] : [],
+      ];
+      return {
+        url,
+        maxDepth: this.params.maxDepth,
+        allowedDomains,
+        priority: pageUrls.length - index,
+      };
     });
+    await crawler.queue(queue);
     await crawler.onIdle(); // Resolved when no queue is left
     await crawler.close(); // Close the crawler
   }