+++ packages/bodiless-migration-tool/src/site-flattener.ts
@@ -42,6 +42,7 @@ import {
   ComponentScope,
   HtmlToComponentsSettings,
 } from './html-to-components';
+import page404Handler, { Page404Params } from './page404-handler';
 import debug from './debug';
 
 export enum TrailingSlash {
@@ -54,6 +55,7 @@ export enum TransformerRule {
   Replace = 'replace',
   ReplaceString = 'replaceString',
   ToComponent = 'tocomponent',
+  RemoveAttribute = 'removeAttribute',
 }
 
 export interface Transformer {
@@ -61,7 +63,8 @@ export interface Transformer {
   selector: string,
   replacement: string,
   context: string,
-  scope?: ComponentScope
+  scope?: ComponentScope,
+  attributes: string[],
 }
 
 export interface SiteFlattenerParams {
@@ -70,6 +73,7 @@ export interface SiteFlattenerParams {
   gitRepository?: string,
   trailingSlash?: TrailingSlash,
   scraperParams: ScraperParams,
+  page404Params: Page404Params,
   steps: {
     setup: boolean,
     scrape: boolean,
@@ -81,6 +85,7 @@ export interface SiteFlattenerParams {
   htmltojsx: boolean,
   useSourceHtml?: boolean,
   disableTailwind?: boolean,
+  reservedPaths?: Array<string>,
   allowFallbackHtml?: boolean,
 }
 
@@ -91,14 +96,16 @@ export class SiteFlattener {
 
   constructor(params: SiteFlattenerParams) {
     this.params = {
+      reservedPaths: [],
       ...params,
       websiteUrl: prependProtocolToBareUrl(params.websiteUrl),
       trailingSlash: params.trailingSlash || TrailingSlash.Add,
       scraperParams: {
         ...params.scraperParams,
-        pageUrl: prependProtocolToBareUrl(params.scraperParams.pageUrl),
+        pageUrls: params.scraperParams.pageUrls.map(pageUrl => prependProtocolToBareUrl(pageUrl)),
       },
     };
+
     const jamStackAppParams: JamStackAppParams = {
       gitRepository: this.params.gitRepository,
       workDir: this.params.workDir,
@@ -131,11 +138,13 @@ export class SiteFlattener {
       enableFileDownload: false,
       downloadPath: getUrlToLocalDirectoryMapper(this.canvasX.getStaticDir()),
     };
+    const { page404Params } = this.params;
     const scraper = new Scraper(scraperParams);
-    scraper.on('success', async result => {
+    scraper.on('pageReceived', async result => {
       try {
-        debug(`scraped url ${result.pageUrl}.`);
-        const pageCreator = new PageCreator(this.getPageCreatorParams(result));
+        debug(`scraped page from ${result.pageUrl}.`);
+        const processedResult = page404Handler.processScrapedPage(result, page404Params);
+        const pageCreator = new PageCreator(this.getPageCreatorParams(processedResult));
         debug(`creating page for ${result.pageUrl}.`);
         await pageCreator.createPage();
       } catch (error) {
@@ -146,11 +155,15 @@ export class SiteFlattener {
       debug(error.message);
     });
     scraper.on('fileReceived', async fileUrl => {
-      const downloader = new Downloader(this.params.websiteUrl, this.canvasX.getStaticDir());
+      const downloader = new Downloader(
+        this.params.websiteUrl, this.canvasX.getStaticDir(), this.params.reservedPaths,
+      );
       await downloader.downloadFiles([fileUrl]);
     });
     scraper.on('requestStarted', async fileUrl => {
-      const downloader = new Downloader(this.params.websiteUrl, this.canvasX.getStaticDir());
+      const downloader = new Downloader(
+        this.params.websiteUrl, this.canvasX.getStaticDir(), this.params.reservedPaths,
+      );
       await downloader.downloadFiles([fileUrl]);
     });
     await scraper.Crawl();
@@ -212,6 +225,37 @@ export class SiteFlattener {
         )
         .forEach(item => htmlParser.replace(item.selector, item.replacement));
     }
+
+    // Cleanup primary attributes to avoid build issue from Helmet.
+    const emptyAttributeRemovalRules = [
+      {
+        selector: 'head link',
+        attributes: ['rel', 'href'],
+      },
+      {
+        selector: 'head meta',
+        attributes: ['name', 'charset', 'http-equiv', 'property', 'itemprop'],
+      },
+      {
+        selector: 'head noscript',
+        attributes: ['innerhtml'],
+      },
+      {
+        selector: 'head link',
+        attributes: ['rel', 'href'],
+      },
+      {
+        selector: 'head script',
+        attributes: ['src', 'innerhtml'],
+      },
+      {
+        selector: 'head style',
+        attributes: ['csstext'],
+      },
+    ];
+    emptyAttributeRemovalRules.forEach(item => {
+      htmlParser.removeEmptyAttribute(item.selector, item.attributes);
+    });
     const pageHtml = htmlParser.getPageHtml();
     return this.transformAttributes(pageHtml);
   }
@@ -229,9 +273,9 @@ export class SiteFlattener {
   }
 
   private getHtmlToComponentsSettings(): HtmlToComponentsSettings {
-    const tranfomers = this.params.transformers || [];
+    const tranformers = this.params.transformers || [];
     const settings: HtmlToComponentsSettings = {
-      rules: tranfomers
+      rules: tranformers
         .filter(item => item.rule === TransformerRule.ToComponent)
         .map(item => ({
           selector: item.selector,
@@ -269,6 +313,7 @@ export class SiteFlattener {
       htmlToComponents: this.params.htmltojsx,
       allowFallbackHtml: this.params.allowFallbackHtml,
       htmlToComponentsSettings,
+      reservedPaths: this.params.reservedPaths,
     };
     return pageCreatorParams;
   }