+++ packages/bodiless-migration-tool/tests/html-parser.test.ts
@@ -344,6 +344,16 @@ describe('transforming html element text to attribute', () => {
     const result = htmlParser.getBodyHtml();
     expect(result).toBe(expected);
   });
+  describe('when html element text is empty', () => {
+    it('should not create an attribute with empty value', () => {
+      const input = '<body><script></script></body>';
+      const expected = '<script></script>';
+      const htmlParser = new HtmlParser(input);
+      htmlParser.transformElementTextToAttribute('script', 'innerHTML');
+      const result = htmlParser.getBodyHtml();
+      expect(result).toBe(expected);
+    });
+  });
 });
 
 describe('getting html and body tags', () => {
@@ -579,6 +589,33 @@ describe('replacing strings in page source html', () => {
     const htmlParser = new HtmlParser(sourceHtml);
     htmlParser.replaceString('" ">', '">');
     const processedHtml = htmlParser.getBodyHtml();
+    expect(htmlclean(processedHtml)).toBe(htmlclean(expectedHtml));
+  });
+});
+
+describe('remove empty attributes on selected element', () => {
+  it('removes empty attributes', () => {
+    const sourceHtml = `
+      <html>
+        <head>
+          <link href="" rel="test">
+        </head>
+        <body>
+        </body>
+      </html>
+   `;
+    const expectedHtml = `
+      <html>
+        <head>
+          <link rel="test">
+        </head>
+        <body>
+        </body>
+      </html>
+    `;
+    const htmlParser = new HtmlParser(sourceHtml);
+    htmlParser.removeEmptyAttribute('head link', ['href']);
+    const processedHtml = htmlParser.getPageHtml();
     expect(htmlclean(processedHtml)).toBe(htmlclean(expectedHtml));
   });
 });